1.add buildpepiline ci and cache server

+++ Assets/Code/BDFramework/Editor/CI/BuildPipeline/BuildPipeLine_CI.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Runtime.Serialization.Json;
 using BDFramework.Editor.Asset;
 using BDFramework.Editor.EditorLife;
+using Code.BDFramework.Core.Tools;
 using Code.BDFramework.Editor;
 using LitJson;
 using UnityEditor;
@@ -13,6 +15,81 @@ using UnityEngine;
 
 namespace BDFramework.Editor
 {
+    public class Response
+    {
+        /// <summary>
+        /// 返回码
+        /// 0=失败
+        /// 1=成功
+        /// </summary>
+        public int Code { get; set; }
+
+        /// <summary>
+        /// 返回的消息
+        /// </summary>
+        public string Msg { get; set; }
+
+        /// <summary>
+        /// 返回的内容
+        /// </summary>
+        public JsonData Content { get; set; }
+
+        public bool IsObject { get; set; } = false;
+
+        //隐藏构造函数
+        private Response()
+        {
+        }
+
+        /// <summary>
+        /// 创建response
+        /// </summary>
+        /// <param name="json"></param>
+        /// <returns></returns>
+        static public Response Cretate(string json)
+        {
+            var ret = new Response();
+            var jw  = JsonMapper.ToObject(json);
+
+            //code
+            int code = -1;
+            int.TryParse(jw["code"].ToString(), out code);
+            ret.Code = code;
+            //msg
+            var msg = jw["msg"];
+            if (msg != null)
+            {
+                ret.Msg = msg.ToJson();
+            }
+
+            //content
+            ret.Content = jw["content"];
+            return ret;
+        }
+
+        /// <summary>
+        /// 获取返回内容
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public T GetContent<T>()
+        {
+            if (this.Content.IsObject || this.Content.IsArray)
+            {
+                return JsonMapper.ToObject<T>(this.Content.ToJson());
+            }
+            else
+            {
+                var type      = typeof(T);
+                var targetObj = Convert.ChangeType(Content.ToString(), type);
+
+                return (T) targetObj;
+            }
+
+            return (T)new object();
+        }
+    }
+
     static public class BuildPipeLine_CI
     {
         private static string outputPath = "";
@@ -25,38 +102,33 @@ namespace BDFramework.Editor
             outputPath = Application.streamingAssetsPath;
         }
 
-        /// <summary>
-        /// 获取平台
-        /// </summary>
-        /// <param name="platform"></param>
-        public static string GetPlatform(RuntimePlatform platform)
-        {
-            if (platform == RuntimePlatform.IPhonePlayer)
-            {
-                return "iOS";
-            }
-            else
-            {
-                return "Android";
-            }
-        }
 
         #region 构建资源
 
         /// <summary>
         /// 构建iOS
         /// </summary>
-        public static void BuildAssetBundle_iOS(string version = "")
+        public static void BuildAssetBundle_iOS()
         {
+            //下载
+            DownloadFormFileServer(RuntimePlatform.IPhonePlayer);
+            //构建
             BuildAssetBundle(RuntimePlatform.IPhonePlayer, BuildTarget.iOS);
+            //上传
+            UploadFormFileServer(RuntimePlatform.IPhonePlayer);
         }
 
         /// <summary>
         /// 构建Android
         /// </summary>
-        public static void BuildAssetBundle_Android(string version = "")
+        public static void BuildAssetBundle_Android()
         {
+            //下载
+            DownloadFormFileServer(RuntimePlatform.Android);
+            //构建
             BuildAssetBundle(RuntimePlatform.Android, BuildTarget.Android);
+            //上传
+            UploadFormFileServer(RuntimePlatform.Android);
         }
 
         /// <summary>
@@ -69,8 +141,6 @@ namespace BDFramework.Editor
             //2.打包模式
             var config = BDFrameEditorConfigHelper.EditorConfig.BuildAssetConfig;
             AssetBundleEditorToolsV2.GenAssetBundle(outputPath, platform, target, BuildAssetBundleOptions.ChunkBasedCompression, true, config.AESCode);
-            //
-            UploadFormFileServer(platform);
         }
 
         #endregion
@@ -80,9 +150,9 @@ namespace BDFramework.Editor
         /// <summary>
         /// 构建dll
         /// </summary>
-        public static void BuildDLL(RuntimePlatform platform)
+        public static void BuildDLL()
         {
-            EditorWindow_ScriptBuildDll.RoslynBuild(outputPath, platform, ScriptBuildTools.BuildMode.Release);
+            EditorWindow_ScriptBuildDll.RoslynBuild(outputPath, RuntimePlatform.Android, ScriptBuildTools.BuildMode.Release);
         }
 
         #endregion
@@ -90,7 +160,7 @@ namespace BDFramework.Editor
         #region 发布资源版本
 
         #endregion
-        
+
         #region 第一次构建包体
 
         static public void BuildAndroidDebug()
@@ -149,21 +219,25 @@ namespace BDFramework.Editor
             {
                 EditorBuildPackage.BuildIpa();
             }
+
+            //最后上传
+            UploadFormFileServer(platform);
         }
 
         #endregion
-        
+
         #region 增量构建包体
 
         #endregion
 
         #region 上传下载美术资源
 
-        public enum FileProtocol
+        public enum ABServer_Protocol
         {
-            GetAllFile,
-            GetFile,
+            GetLastUploadFiles,
             Upload,
+            Download,
+            GetLastVersion
         }
 
 
@@ -173,32 +247,53 @@ namespace BDFramework.Editor
         /// <param name="platform"></param>
         private static bool DownloadFormFileServer(RuntimePlatform platform)
         {
-            var platformStr = GetPlatform(platform);
-            var url         = BDFrameEditorConfigHelper.EditorConfig.BuildAssetConfig.AssetBundleFileServerUrl;
-            var protocol    = string.Format("{0}/{1}/{2}", url, nameof(FileProtocol.GetAllFile), platformStr);
+            var platformStr = BDApplication.GetPlatformPath(platform);
+            var url         = BDFrameEditorConfigHelper.EditorConfig.BuildAssetConfig.AssetBundleFileServerUrl + "/Assetbundle";
             var webclient   = new WebClient();
+
+            //获取最新版本的文件 //url + 协议 +参数 
+            var protocol = string.Format("{0}/{1}/{2}/{3}", url, nameof(ABServer_Protocol.GetLastUploadFiles), PlayerSettings.applicationIdentifier, platformStr);
+
             //所有文件的配置
-            string fileConfig = "";
+            string ret = "";
             try
             {
-                fileConfig = webclient.DownloadString(protocol);
+                ret = webclient.DownloadString(protocol);
             }
             catch (Exception e)
             {
-                Debug.LogError("服务器文件获取失败、或者无服务");
+                Debug.LogError(e.Data);
                 return true;
+            }
+
+            //获取返回的文件列表
+            var response = Response.Cretate(ret);
+            if (response.Code == 0)
+            {
+                Debug.LogError(response.Msg);
+                return false;
+            }
+            var fs        = response.GetContent<List<string>>();
+            var fileQueue = new Queue<string>();
+            foreach (var f in fs)
+            {
+                fileQueue.Enqueue(f);
             }
 
             //删除本地所有素材
             var localPath = string.Format("{0}/{1}/Art", outputPath, platformStr);
-            Directory.Delete(localPath, true);
+            if (Directory.Exists(localPath))
+            {
+                Directory.Delete(localPath, true);
+            }
+
+            Directory.CreateDirectory(localPath);
             //
-            var fileQueue = JsonMapper.ToObject<Queue<string>>(fileConfig);
             int ErrorCont = 0;
             while (fileQueue.Count > 0)
             {
                 var filename = fileQueue.Dequeue();
-                var furl     = string.Format("{0}/{1}/{2}/{3}", url, nameof(FileProtocol.GetFile), platformStr, filename);
+                var furl     = string.Format("{0}/{1}/{2}/{3}/{4}", url, nameof(ABServer_Protocol.Download), PlayerSettings.applicationIdentifier, platformStr, filename);
                 var savePath = string.Format("{0}/{1}", localPath, filename);
 
                 try
@@ -210,15 +305,16 @@ namespace BDFramework.Editor
                     //重新压栈执行
                     fileQueue.Enqueue(filename);
                     ErrorCont++;
-                    if (ErrorCont > 1000)
+                    if (ErrorCont > 100)
                     {
-                        Debug.LogError("失败次数过多，本次任务失败");
+                        Debug.LogError("失败次数过多，下载失败:" + e.Message);
                         return false;
                     }
                 }
             }
 
             webclient.Dispose();
+            Debug.Log("所有ab下载成功!");
             return true;
         }
 
@@ -228,36 +324,65 @@ namespace BDFramework.Editor
         /// <param name="platform"></param>
         private static void UploadFormFileServer(RuntimePlatform platform)
         {
-            var platformStr = GetPlatform(platform);
-            var url         = BDFrameEditorConfigHelper.EditorConfig.BuildAssetConfig.AssetBundleFileServerUrl;
-            var protocol    = string.Format("{0}/{1}/{2}", url, nameof(FileProtocol.Upload), platformStr);
-            var webclient   = new WebClient();
+            var platformStr = BDApplication.GetPlatformPath(platform);
+            var url         = BDFrameEditorConfigHelper.EditorConfig.BuildAssetConfig.AssetBundleFileServerUrl + "/Assetbundle";
+            ;
+            var webclient = new WebClient();
+            //获取版本号
+            var protocol = string.Format("{0}/{1}/{2}/{3}", url, nameof(ABServer_Protocol.GetLastVersion), PlayerSettings.applicationIdentifier, platformStr);
+            int version  = 0;
+            try
+            {
+                var ret      = webclient.DownloadString(protocol);
+                var response = Response.Cretate(ret);
+                version = response.GetContent<int>();
+                version++;
+            }
+            catch (Exception e)
+            {
+                Debug.LogError("版本号获取错误:" + e.Data);
+            }
+
+            //上传
+            protocol = string.Format("{0}/{1}/{2}/{3}/{4}", url, nameof(ABServer_Protocol.Upload), PlayerSettings.applicationIdentifier, version, platformStr);
             //获取本地所有素材
             var localPath = string.Format("{0}/{1}/Art", outputPath, platformStr);
-            var fs        = Directory.GetFiles(localPath, "*.*", SearchOption.AllDirectories).Where((f) => !f.EndsWith(".meta"));
-            var fsQueue   = new Queue<string>(fs);
+            if (!Directory.Exists(localPath))
+            {
+                Directory.CreateDirectory(localPath);
+            }
+
+            var fs = Directory.GetFiles(localPath, "*", SearchOption.AllDirectories).Where((f) => !f.EndsWith(".meta")).ToList();
             //
             int ErrorCont = 0;
-            foreach (var f in fsQueue)
+            for (int i = 0; i < fs.Count; i++)
             {
+                var f = fs[i];
+
                 try
                 {
                     webclient.UploadFile(protocol, f);
                 }
                 catch (Exception e)
                 {
-                    fsQueue.Enqueue(f);
                     //重新压栈执行
+                    i--;
+                    fs.RemoveAt(0);
+                    fs.Add(f);
+                    //错误次数
                     ErrorCont++;
                     if (ErrorCont > 10)
                     {
-                        Debug.LogError("失败次数过多，本次任务失败");
+                        Debug.LogError(e.Message);
+                        Debug.LogError("失败次数过多，本次任务失败:" + protocol);
                         return;
                     }
                 }
             }
 
             webclient.Dispose();
+
+            Debug.Log("所有ab上传成功,版本号:" + version);
         }
 
         #endregion
